Move validation logic for shiro to ShiroSecurityService
